Move require_aligned to misc.py
