+++ disruption_py/core/utils/misc.py
@@ -17,6 +17,8 @@ from typing import Dict, List, Tuple, Type
 import numpy as np
 from loguru import logger
 
+from disruption_py.core.physics_method.errors import CalculationError
+
 
 def without_duplicates(lst: List):
     """
@@ -207,6 +209,39 @@ def to_tuple(
     Dict[str, Tuple[str, np.ndarray]]
     """
     return {k: (dim, v) for k, v in data.items()}
+
+
+def require_aligned(path: str, signal: np.ndarray, time: np.ndarray) -> np.ndarray:
+    """
+    Check a signal is sampled on the time base it will be interpolated against.
+
+    The last axis is the time axis for both 1-D traces and 2-D (channel, time)
+    profiles, so this covers both.
+
+    Parameters
+    ----------
+    path : str
+        Path of the signal, used for the error message.
+    signal : np.ndarray
+        The signal to check.
+    time : np.ndarray
+        The time base the signal should be sampled on.
+
+    Returns
+    -------
+    np.ndarray
+        The signal, unchanged.
+
+    Raises
+    ------
+    CalculationError
+        If the signal's time axis does not match the length of the time base.
+    """
+    if signal.shape[-1] != len(time):
+        raise CalculationError(
+            f"{path} has {signal.shape[-1]} samples but its time base has {len(time)}"
+        )
+    return signal
 
 
 def filter_dict(i: Dict, s: str) -> Dict:
